accessible/tests/mochitest/tree/test_txtctrl.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL textbox and textarea hierarchy tests">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../events.js" />

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Test

    function doTest()
    {
      ////////////////////
      // textbox
      ////////////////////
      var accTree = {
        role: ROLE_ENTRY,
        children: [
          {
            role: ROLE_TEXT_LEAF,
            children: []
          }
        ]
      };

      // default textbox
      testAccessibleTree("txc1", accTree);

      // number textbox
      testAccessibleTree("txc2", accTree);

      // search textbox
      testAccessibleTree("txc3", accTree);

      // timed textbox
      testAccessibleTree("txc4_deprecated", accTree);

      ////////////////////
      // password textbox
      ////////////////////
      accTree = {
        role: ROLE_PASSWORD_TEXT,
        children: [
          {
            role: ROLE_TEXT_LEAF,
            children: []
          }
        ]
      };

      testAccessibleTree("txc5", accTree);

      ////////////////////
      // multiline textbox
      ////////////////////
      accTree = {
        role: ROLE_ENTRY,
        children: [
          {
            role: ROLE_TEXT_LEAF,
            children: []
          }
        ]
      };

      testAccessibleTree("txc6", accTree);

      ////////////////////
      // autocomplete textbox
      ////////////////////
      accTree = {
        // textbox
        role: ROLE_AUTOCOMPLETE,
        children: [
          {
            // html:input
            role: ROLE_ENTRY,
            children: [
              {
                // #text
                role: ROLE_TEXT_LEAF,
                children: []
              }
            ]
          },
          {
            // xul:menupopup
            role: ROLE_COMBOBOX_LIST,
            children: []
          }
        ]
      };

      function test_txc7() {
        testAccessibleTree("txc7", accTree);
        SimpleTest.finish();
      }

      // XPFE and Toolkit autocomplete widgets differ.
      var txc7 = document.getElementById("txc7");
      if ("clearResults" in txc7) {
        SimpleTest.ok(true, "Testing (Old) XPFE autocomplete widget.");

        // Popup is always created. (See code below.)

        accTree.children.push(
          {
            // xul:panel
            role: ROLE_COMBOBOX_LIST,
            children: [
              {
                // xul:tree
                role: ROLE_TABLE,
                children: [
                  {
                    // xul:treecols
                    role: ROLE_LIST,
                    children: [
                      {
                        // xul:treecol
                        role: ROLE_COLUMNHEADER,
                        children: []
                      }
                    ]
                  }
                ]
              }
            ]
          }
        );
        test_txc7();

      } else {
        SimpleTest.ok(true, "Testing (New) Toolkit autocomplete widget.");

        // Dumb access to trigger popup lazy creation.
        waitForEvent(EVENT_REORDER, txc7, test_txc7);
        txc7.popup;

        accTree.children.push(
          {
            role: ROLE_LIST,
            children: [
              {
                role: ROLE_LIST,
                children: [
                  {
                    role: ROLE_COLUMNHEADER,
                    children: []
                  }
                ]
              }
            ]
          }
        );
      }
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=542824"
         title="Create child accessibles for text controls from native anonymous content">
        Mozilla Bug 542824
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <textbox id="txc1" value="hello"/>
      <textbox id="txc2" type="number" value="44"/>
      <textbox id="txc3" type="search" value="hello"/>
      <!-- This textbox triggers "Warning: Timed textboxes are deprecated. Consider using type="search" instead.".
           Yet let's test it too as long as it's (still) supported. -->
      <textbox id="txc4_deprecated" type="timed" value="hello"/>
      <textbox id="txc5" type="password" value="hello"/>
      <textbox id="txc6" multiline="true" value="hello"/>
      <textbox id="txc7" type="autocomplete" value="hello"/>
    </vbox>
  </hbox>

</window>
